@@ -20,192 +20,191 @@
  * SOFTWARE.
  */
 
-namespace TBC.OpenBanking.Jws
+namespace TBC.OpenBanking.Jws;
+
+using System;
+using System.Collections.Generic;
+using System.Globalization;
+using System.Security.Cryptography;
+using System.Security.Cryptography.X509Certificates;
+using System.Text;
+
+/// <summary>
+/// Table is based on <see href="https://tools.ietf.org/html/rfc7518#section-3.1"/>
+/// <code>
+/// <para>+--------------+-------------------------------+--------------------+</para>
+/// <para>| "alg" Param  | Digital Signature or MAC      | Implementation     |</para>
+/// <para>| Value        | Algorithm                     | Requirements       |</para>
+/// <para>+--------------+-------------------------------+--------------------+</para>
+/// <para>| RS256        | RSASSA-PKCS1-v1_5 using       | Recommended        |</para>
+/// <para>|              | SHA-256                       |                    |</para>
+/// <para>| RS384        | RSASSA-PKCS1-v1_5 using       | Optional           |</para>
+/// <para>|              | SHA-384                       |                    |</para>
+/// <para>| RS512        | RSASSA-PKCS1-v1_5 using       | Optional           |</para>
+/// <para>|              | SHA-512                       |                    |</para>
+/// <para>| PS256        | RSASSA-PSS using SHA-256 and  | Optional           |</para>
+/// <para>|              | MGF1 with SHA-256             |                    |</para>
+/// <para>| PS384        | RSASSA-PSS using SHA-384 and  | Optional           |</para>
+/// <para>|              | MGF1 with SHA-384             |                    |</para>
+/// <para>| PS512        | RSASSA-PSS using SHA-512 and  | Optional           |</para>
+/// <para>|              | MGF1 with SHA-512             |                    |</para>
+/// <para>+--------------+-------------------------------+--------------------+</para>
+/// </code>
+/// https://tools.ietf.org/html/rfc3447#section-8.2
+/// </summary>
+public class AlgorithmRsaSsa : Algorithm
 {
-    using System;
-    using System.Collections.Generic;
-    using System.Globalization;
-    using System.Security.Cryptography;
-    using System.Security.Cryptography.X509Certificates;
-    using System.Text;
+    private RSA rsaPrivate;
+    private RSA rsaPublic;
+    private HashAlgorithmName hashName;
+    private RSASignaturePadding padding;
+    private string algorithName;
+
+    /// <inheritdoc/>
+    public override string Name => algorithName;
+
+    /// <inheritdoc/>
+    public override HashAlgorithmName HashAlgorithmName => hashName;
+
+    public AlgorithmRsaSsa(int keySize, HashAlgorithmName hashName, RSASignaturePadding padding)
+    {
+        var r = RSA.Create();
+        r.KeySize = keySize;
+
+        Init(r, r, hashName, padding);
+    }
+
+    public AlgorithmRsaSsa(X509Certificate2 cert, HashAlgorithmName hashName, RSASignaturePadding padding)
+    {
+        if (cert == null) throw new ArgumentNullException(nameof(cert));
+        if (padding == null) throw new ArgumentNullException(nameof(padding));
+
+        var privateKey = cert.GetRSAPrivateKey();
+        var publicKey = cert.GetRSAPublicKey();
+
+        Init(privateKey, publicKey, hashName, padding);
+    }
+
+    public AlgorithmRsaSsa(RSA r, HashAlgorithmName hashName, RSASignaturePadding padding)
+    {
+        // TODO: The right way is to check if RSA contains private and public and set rsaPrivate and rsaPublic accordingly
+        Init(r, r, hashName, padding);
+    }
+
+    public AlgorithmRsaSsa(RSAParameters Parameters, HashAlgorithmName hashName, RSASignaturePadding padding)
+    {
+        var r = RSA.Create();
+        r.ImportParameters(Parameters);
+
+        // TODO: The right way is to check if parameters contains private and public and set rsaPrivate and rsaPublic accordingly
+        Init(r, r, hashName, padding);
+    }
+
+    ///// <summary>
+    ///// RSA Cryptographic service provider.
+    ///// </summary>
+    //public RSA AsymmetricAlgorithm => rsaPrivate;
 
     /// <summary>
-    /// Table is based on <see href="https://tools.ietf.org/html/rfc7518#section-3.1"/>
-    /// <code>
-    /// <para>+--------------+-------------------------------+--------------------+</para>
-    /// <para>| "alg" Param  | Digital Signature or MAC      | Implementation     |</para>
-    /// <para>| Value        | Algorithm                     | Requirements       |</para>
-    /// <para>+--------------+-------------------------------+--------------------+</para>
-    /// <para>| RS256        | RSASSA-PKCS1-v1_5 using       | Recommended        |</para>
-    /// <para>|              | SHA-256                       |                    |</para>
-    /// <para>| RS384        | RSASSA-PKCS1-v1_5 using       | Optional           |</para>
-    /// <para>|              | SHA-384                       |                    |</para>
-    /// <para>| RS512        | RSASSA-PKCS1-v1_5 using       | Optional           |</para>
-    /// <para>|              | SHA-512                       |                    |</para>
-    /// <para>| PS256        | RSASSA-PSS using SHA-256 and  | Optional           |</para>
-    /// <para>|              | MGF1 with SHA-256             |                    |</para>
-    /// <para>| PS384        | RSASSA-PSS using SHA-384 and  | Optional           |</para>
-    /// <para>|              | MGF1 with SHA-384             |                    |</para>
-    /// <para>| PS512        | RSASSA-PSS using SHA-512 and  | Optional           |</para>
-    /// <para>|              | MGF1 with SHA-512             |                    |</para>
-    /// <para>+--------------+-------------------------------+--------------------+</para>
-    /// </code>
-    /// https://tools.ietf.org/html/rfc3447#section-8.2
+    /// Signs data.
     /// </summary>
-    public class AlgorithmRsaSsa : Algorithm
+    /// <param name="headerEncoded">Encoded properties to include in the header.</param>
+    /// <param name="payloadEncoded">Encoded properties to include in the payload.</param>
+    /// <returns>Signature encoded as Base64Url string</returns>
+    public override string Sign(string headerEncoded, string payloadEncoded)
     {
-        private RSA rsaPrivate;
-        private RSA rsaPublic;
-        private HashAlgorithmName hashName;
-        private RSASignaturePadding padding;
-        private string algorithName;
+        if (rsaPrivate == null) throw new CryptographicException("Private key is not set");
 
-        /// <inheritdoc/>
-        public override string Name => algorithName;
+        byte[] data = Encoding.ASCII.GetBytes(headerEncoded + "." + payloadEncoded);
+        byte[] signature = SignData(data);
 
-        /// <inheritdoc/>
-        public override HashAlgorithmName HashAlgorithmName => hashName;
+        return signature.EncodeBase64Url();
+    }
 
-        public AlgorithmRsaSsa(int keySize, HashAlgorithmName hashName, RSASignaturePadding padding)
+    /// <inheritdoc/>
+    public override byte[] SignData(byte[] data)
+    {
+        if (rsaPrivate == null) throw new CryptographicException("Private key is not set");
+
+        byte[] signature;
+        lock (rsaPrivate)
         {
-            var r = RSA.Create();
-            r.KeySize = keySize;
-
-            Init(r, r, hashName, padding);
+            signature = rsaPrivate.SignData(data, hashName, padding);
         }
 
-        public AlgorithmRsaSsa(X509Certificate2 cert, HashAlgorithmName hashName, RSASignaturePadding padding)
+        return signature;
+    }
+
+    /// <inheritdoc/>
+    public override bool VerifySignature(string headerEncoded, string payloadEncoded, string signatureEncoded)
+    {
+        if (rsaPublic == null) throw new CryptographicException("Public key is not set");
+
+        byte[] TokenBin = Encoding.ASCII.GetBytes(headerEncoded + "." + payloadEncoded);
+        return rsaPublic.VerifyData(TokenBin, signatureEncoded.DecodeBase64Url(), hashName, padding);
+    }
+
+    /// <inheritdoc/>
+    public override void Dispose()
+    {
+        rsaPrivate?.Dispose();
+
+        if (rsaPublic != null && rsaPublic != rsaPrivate)
         {
-            if (cert == null) throw new ArgumentNullException(nameof(cert));
-            if (padding == null) throw new ArgumentNullException(nameof(padding));
-
-            var privateKey = cert.GetRSAPrivateKey();
-            var publicKey = cert.GetRSAPublicKey();
-
-            Init(privateKey, publicKey, hashName, padding);
+            rsaPublic.Dispose();
         }
 
-        public AlgorithmRsaSsa(RSA r, HashAlgorithmName hashName, RSASignaturePadding padding)
+        GC.SuppressFinalize(this);
+    }
+
+    public override IDictionary<string, string> GetJwk(bool includePrivate)
+    {
+        throw new NotImplementedException();
+
+        /*
+        RSAParameters parameters = rsaPrivate.ExportParameters(includePrivate);
+
+        var dic = new Dictionary<string, string>(StringComparer.Ordinal)
         {
-            // TODO: The right way is to check if RSA contains private and public and set rsaPrivate and rsaPublic accordingly
-            Init(r, r, hashName, padding);
+            { "kty", "RSA" },
+            { "n", parameters.Modulus.EncodeBase64Url() },
+            { "e", parameters.Exponent.EncodeBase64Url() },
+        };
+
+        if (includePrivate)
+        {
+            dic.Add("d", Base64Url.Encode(parameters.D));
+            dic.Add("p", Base64Url.Encode(parameters.P));
+            dic.Add("q", Base64Url.Encode(parameters.Q));
+            dic.Add("dp", Base64Url.Encode(parameters.DP));
+            dic.Add("dq", Base64Url.Encode(parameters.DQ));
+            dic.Add("qi", Base64Url.Encode(parameters.InverseQ));
         }
 
-        public AlgorithmRsaSsa(RSAParameters Parameters, HashAlgorithmName hashName, RSASignaturePadding padding)
-        {
-            var r = RSA.Create();
-            r.ImportParameters(Parameters);
+        return dic;
+        */
+    }
 
-            // TODO: The right way is to check if parameters contains private and public and set rsaPrivate and rsaPublic accordingly
-            Init(r, r, hashName, padding);
+    private void Init(RSA privateKey, RSA publicKey, HashAlgorithmName hashName, RSASignaturePadding padding)
+    {
+        this.rsaPrivate = privateKey;
+        this.rsaPublic = publicKey;
+        this.hashName = hashName;
+        this.padding = padding;
+        this.algorithName = CreateAlgorithmName(padding, hashName);
+    }
+
+    private string CreateAlgorithmName(RSASignaturePadding padding, HashAlgorithmName hashName)
+    {
+        int hashSize;
+        if (hashName == HashAlgorithmName.SHA256) hashSize = 256;
+        else if (hashName == HashAlgorithmName.SHA384) hashSize = 386;
+        else if (hashName == HashAlgorithmName.SHA512) hashSize = 512;
+        else
+        {
+            throw new CryptographicException($"Unsuitable hash algorithm '{hashName.Name}'.");
         }
 
-        ///// <summary>
-        ///// RSA Cryptographic service provider.
-        ///// </summary>
-        //public RSA AsymmetricAlgorithm => rsaPrivate;
-
-        /// <summary>
-        /// Signs data.
-        /// </summary>
-        /// <param name="headerEncoded">Encoded properties to include in the header.</param>
-        /// <param name="payloadEncoded">Encoded properties to include in the payload.</param>
-        /// <returns>Signature encoded as Base64Url string</returns>
-        public override string Sign(string headerEncoded, string payloadEncoded)
-        {
-            if (rsaPrivate == null) throw new CryptographicException("Private key is not set");
-
-            byte[] data = Encoding.ASCII.GetBytes(headerEncoded + "." + payloadEncoded);
-            byte[] signature = SignData(data);
-
-            return signature.EncodeBase64Url();
-        }
-
-        /// <inheritdoc/>
-        public override byte[] SignData(byte[] data)
-        {
-            if (rsaPrivate == null) throw new CryptographicException("Private key is not set");
-
-            byte[] signature;
-            lock (rsaPrivate)
-            {
-                signature = rsaPrivate.SignData(data, hashName, padding);
-            }
-
-            return signature;
-        }
-
-        /// <inheritdoc/>
-        public override bool VerifySignature(string headerEncoded, string payloadEncoded, string signatureEncoded)
-        {
-            if (rsaPublic == null) throw new CryptographicException("Public key is not set");
-
-            byte[] TokenBin = Encoding.ASCII.GetBytes(headerEncoded + "." + payloadEncoded);
-            return rsaPublic.VerifyData(TokenBin, signatureEncoded.DecodeBase64Url(), hashName, padding);
-        }
-
-        /// <inheritdoc/>
-        public override void Dispose()
-        {
-            rsaPrivate?.Dispose();
-
-            if (rsaPublic != null && rsaPublic != rsaPrivate)
-            {
-                rsaPublic.Dispose();
-            }
-
-            GC.SuppressFinalize(this);
-        }
-
-        public override IDictionary<string, string> GetJwk(bool includePrivate)
-        {
-            throw new NotImplementedException();
-
-            /*
-            RSAParameters parameters = rsaPrivate.ExportParameters(includePrivate);
-
-            var dic = new Dictionary<string, string>(StringComparer.Ordinal)
-            {
-                { "kty", "RSA" },
-                { "n", parameters.Modulus.EncodeBase64Url() },
-                { "e", parameters.Exponent.EncodeBase64Url() },
-            };
-
-            if (includePrivate)
-            {
-                dic.Add("d", Base64Url.Encode(parameters.D));
-                dic.Add("p", Base64Url.Encode(parameters.P));
-                dic.Add("q", Base64Url.Encode(parameters.Q));
-                dic.Add("dp", Base64Url.Encode(parameters.DP));
-                dic.Add("dq", Base64Url.Encode(parameters.DQ));
-                dic.Add("qi", Base64Url.Encode(parameters.InverseQ));
-            }
-
-            return dic;
-            */
-        }
-
-        private void Init(RSA privateKey, RSA publicKey, HashAlgorithmName hashName, RSASignaturePadding padding)
-        {
-            this.rsaPrivate = privateKey;
-            this.rsaPublic = publicKey;
-            this.hashName = hashName;
-            this.padding = padding;
-            this.algorithName = CreateAlgorithmName(padding, hashName);
-        }
-
-        private string CreateAlgorithmName(RSASignaturePadding padding, HashAlgorithmName hashName)
-        {
-            int hashSize;
-            if (hashName == HashAlgorithmName.SHA256) hashSize = 256;
-            else if (hashName == HashAlgorithmName.SHA384) hashSize = 386;
-            else if (hashName == HashAlgorithmName.SHA512) hashSize = 512;
-            else
-            {
-                throw new CryptographicException($"Unsuitable hash algorithm '{hashName.Name}'.");
-            }
-
-            return ((padding == RSASignaturePadding.Pkcs1) ? "RS" : "PS") + hashSize.ToString(CultureInfo.InvariantCulture);
-        }
+        return ((padding == RSASignaturePadding.Pkcs1) ? "RS" : "PS") + hashSize.ToString(CultureInfo.InvariantCulture);
     }
 }
