@@ -20,88 +20,87 @@
  * SOFTWARE.
  */
 
-namespace TBC.OpenBanking.Jws
+namespace TBC.OpenBanking.Jws;
+
+using System;
+using System.Collections.Generic;
+using System.Security.Cryptography;
+
+internal class HttpDigest
 {
-    using System;
-    using System.Collections.Generic;
-    using System.Security.Cryptography;
+    private readonly HashAlgorithm hashAlgorithm;
+    private readonly string digestPrefix;
 
-    internal class HttpDigest
+    internal static readonly List<(HashAlgorithmName AlgorithName, string Prefix)> supportedAlgorithms = new(2)
     {
-        private readonly HashAlgorithm hashAlgorithm;
-        private readonly string digestPrefix;
+        (HashAlgorithmName.SHA256, "SHA-256"),
+        (HashAlgorithmName.SHA384, "SHA-384"),
+        (HashAlgorithmName.SHA512, "SHA-512"),
+    };
 
-        internal static readonly List<(HashAlgorithmName AlgorithName, string Prefix)> supportedAlgorithms = new(2)
-        {
-            (HashAlgorithmName.SHA256, "SHA-256"),
-            (HashAlgorithmName.SHA384, "SHA-384"),
-            (HashAlgorithmName.SHA512, "SHA-512"),
-        };
+    internal HashAlgorithm HashAlgorithm => this.hashAlgorithm;
 
-        internal HashAlgorithm HashAlgorithm => this.hashAlgorithm;
+    internal HttpDigest(HashAlgorithmName hashAlgorithmName)
+    {
+        digestPrefix = GetDigestHeader(hashAlgorithmName);
+        hashAlgorithm = HashAlgorithm.Create(hashAlgorithmName.Name);
+    }
 
-        internal HttpDigest(HashAlgorithmName hashAlgorithmName)
-        {
-            digestPrefix = GetDigestHeader(hashAlgorithmName);
-            hashAlgorithm = HashAlgorithm.Create(hashAlgorithmName.Name);
-        }
+    internal static HttpDigest CreateDigest(string digestString)
+    {
+        // Example:
+        // SHA-256=+xeh7JAayYPh8K13UnQCBBcniZzsyat+KDiuy8aZYdI=
 
-        internal static HttpDigest CreateDigest(string digestString)
-        {
-            // Example:
-            // SHA-256=+xeh7JAayYPh8K13UnQCBBcniZzsyat+KDiuy8aZYdI=
+        int dividerIndex = digestString.IndexOf('=');
+        if (dividerIndex == -1)
+            throw new ArgumentOutOfRangeException(nameof(digestString), "Bad format of digest string. Can't find algorithm prefix");
 
-            int dividerIndex = digestString.IndexOf('=');
-            if (dividerIndex == -1)
-                throw new ArgumentOutOfRangeException(nameof(digestString), "Bad format of digest string. Can't find algorithm prefix");
+        var algName = digestString.Substring(0, dividerIndex).Trim();
+        var element = supportedAlgorithms.Find(x => x.Prefix == algName);
+        if (element == default)
+            throw new ArgumentOutOfRangeException(nameof(digestString), "Unsupported hash algorithm");
 
-            var algName = digestString.Substring(0, dividerIndex).Trim();
-            var element = supportedAlgorithms.Find(x => x.Prefix == algName);
-            if (element == default)
-                throw new ArgumentOutOfRangeException(nameof(digestString), "Unsupported hash algorithm");
+        return new HttpDigest(element.AlgorithName);
+    }
 
-            return new HttpDigest(element.AlgorithName);
-        }
+    private string GetDigestHeader(HashAlgorithmName hashAlgorithmName)
+    {
+        var element = supportedAlgorithms.Find(x => x.AlgorithName.Name == hashAlgorithmName.Name);
+        if (element == default)
+            throw new ArgumentOutOfRangeException(nameof(hashAlgorithmName), "Unsupported hash algorithm");
 
-        private string GetDigestHeader(HashAlgorithmName hashAlgorithmName)
-        {
-            var element = supportedAlgorithms.Find(x => x.AlgorithName.Name == hashAlgorithmName.Name);
-            if (element == default)
-                throw new ArgumentOutOfRangeException(nameof(hashAlgorithmName), "Unsupported hash algorithm");
+        return element.Prefix;
+    }
 
-            return element.Prefix;
-        }
+    internal string CalculateDigest(string body)
+    {
+        return CalculateDigest(UTF8EncodingSealed.Instance.GetBytes(body));
+    }
 
-        internal string CalculateDigest(string body)
-        {
-            return CalculateDigest(UTF8EncodingSealed.Instance.GetBytes(body));
-        }
+    internal string CalculateDigest(byte[] body)
+    {
+        if (body is null) throw new ArgumentNullException(nameof(body));
 
-        internal string CalculateDigest(byte[] body)
-        {
-            if (body is null) throw new ArgumentNullException(nameof(body));
-
-            byte[] hashValue = hashAlgorithm.ComputeHash(body);
-            string hashEncoded = Convert.ToBase64String(hashValue);
-            return $"{digestPrefix}={hashEncoded}";
-        }
+        byte[] hashValue = hashAlgorithm.ComputeHash(body);
+        string hashEncoded = Convert.ToBase64String(hashValue);
+        return $"{digestPrefix}={hashEncoded}";
+    }
 
 #if NETCOREAPP2_1_OR_GREATER
-        internal string CalculateDigest(ReadOnlySpan<byte> body)
-        {
-            Span<byte> hash = stackalloc byte[hashAlgorithm.HashSize / 8];
-            hashAlgorithm.TryComputeHash(body, hash, out int bytesWritten);
-            Span<char> chars = stackalloc char[ToBase64_CalculateAndValidateOutputLength(bytesWritten)];
-            Convert.TryToBase64Chars(hash, chars, out _, Base64FormattingOptions.None);
-            return digestPrefix + "=" + new string(chars);
-        }
-
-        private static int ToBase64_CalculateAndValidateOutputLength(int inputLength)
-        {
-            int outlen = inputLength / 3 * 4;
-            outlen += ((inputLength % 3) != 0) ? 4 : 0;
-            return outlen;
-        }
-#endif
+    internal string CalculateDigest(ReadOnlySpan<byte> body)
+    {
+        Span<byte> hash = stackalloc byte[hashAlgorithm.HashSize / 8];
+        hashAlgorithm.TryComputeHash(body, hash, out int bytesWritten);
+        Span<char> chars = stackalloc char[ToBase64_CalculateAndValidateOutputLength(bytesWritten)];
+        Convert.TryToBase64Chars(hash, chars, out _, Base64FormattingOptions.None);
+        return digestPrefix + "=" + new string(chars);
     }
+
+    private static int ToBase64_CalculateAndValidateOutputLength(int inputLength)
+    {
+        int outlen = inputLength / 3 * 4;
+        outlen += ((inputLength % 3) != 0) ? 4 : 0;
+        return outlen;
+    }
+#endif
 }
