@@ -20,111 +20,109 @@
  * SOFTWARE.
  */
 
-namespace TBC.OpenBanking.Jws
+namespace TBC.OpenBanking.Jws;
+
+using System;
+using System.Collections.Generic;
+using System.Text;
+using TBC.OpenBanking.Jws.Exceptions;
+
+/// <summary>
+/// Container for incoming or outgoing HTTP request data.
+/// </summary>
+public class HttpResponseData : HttpMessageData
 {
-    using System;
-    using System.Collections.Generic;
-    using System.Text;
-    using TBC.OpenBanking.Jws.Exceptions;
-
     /// <summary>
-    /// Container for incoming or outgoing HTTP request data.
+    /// (response-status) represents status code
     /// </summary>
-    public class HttpResponseData : HttpMessageData
+    public const string ResponseStatusHeaderName = "(response-status)";
+
+    public readonly static IReadOnlyList<(string Name, HeaderNecessity Necessity)> NecessaryHeaders =
+        new List<(string, HeaderNecessity)>
+        {
+            (ResponseStatusHeaderName, HeaderNecessity.Mandatory),
+            ("x-request-id", HeaderNecessity.Mandatory),
+            ("content-type", HeaderNecessity.IfExists),
+            ("content-length", HeaderNecessity.IfExists),
+            (DigestHeadertName, HeaderNecessity.Mandatory),
+        };
+
+    public string StatusCode { get; set; }
+
+    public override void CheckMandatoryHeaders()
     {
-        /// <summary>
-        /// (response-status) represents status code
-        /// </summary>
-        public const string ResponseStatusHeaderName = "(response-status)";
-
-        public readonly static IReadOnlyList<(string Name, HeaderNecessity Necessity)> NecessaryHeaders =
-            new List<(string, HeaderNecessity)>
-            {
-                (ResponseStatusHeaderName, HeaderNecessity.Mandatory),
-                ("x-request-id", HeaderNecessity.Mandatory),
-                ("content-type", HeaderNecessity.IfExists),
-                ("content-length", HeaderNecessity.IfExists),
-                (DigestHeadertName, HeaderNecessity.Mandatory),
-            };
-
-        public string StatusCode { get; set; }
-
-        public override void CheckMandatoryHeaders()
+        foreach (var (name, necessity) in NecessaryHeaders)
         {
-            foreach (var (name, necessity) in NecessaryHeaders)
-            {
-                if (string.Equals(name, ResponseStatusHeaderName, StringComparison.OrdinalIgnoreCase))
-                    continue;
+            if (string.Equals(name, ResponseStatusHeaderName, StringComparison.OrdinalIgnoreCase))
+                continue;
 
-                if (necessity == HeaderNecessity.Mandatory && !Headers.ContainsKey(name))
-                    throw new HeaderMissingException($"Mandatory header '{name}' is missing");
+            if (necessity == HeaderNecessity.Mandatory && !Headers.ContainsKey(name))
+                throw new HeaderMissingException($"Mandatory header '{name}' is missing");
+        }
+    }
+
+    public override string ComposeHeadersForSignature(IList<string> headers, IDictionary<string, string> additionalHeaders = null)
+    {
+        var sb = new StringBuilder();
+        foreach (var hn in headers)
+        {
+            if (sb.Length > 0)
+                sb.Append(HttpMessageData.HeaderTerminatorInPayload);
+
+            if (string.Equals(hn, ResponseStatusHeaderName, StringComparison.OrdinalIgnoreCase))
+            {
+                sb.Append(hn)
+                    .Append(HttpMessageData.HeaderNameValueSeparator)
+                    .Append(StatusCode);
+            }
+            else
+            {
+                if (!Headers.TryGetValue(hn, out string headerValue)
+                    && (additionalHeaders == null || !additionalHeaders.TryGetValue(hn, out headerValue)))
+                {
+                    throw new HeaderMissingException($"Can't find header '{hn}'");
+                }
+
+                sb.Append(hn)
+                    .Append(HeaderNameValueSeparator)
+                    .Append(headerValue);
             }
         }
 
-        public override string ComposeHeadersForSignature(IList<string> headers, IDictionary<string, string> additionalHeaders = null)
-        {
-            var sb = new StringBuilder();
-            foreach (var hn in headers)
-            {
-                if (sb.Length > 0)
-                    sb.Append(HttpMessageData.HeaderTerminatorInPayload);
+        return sb.ToString();
+    }
 
-                if (string.Equals(hn, ResponseStatusHeaderName, StringComparison.OrdinalIgnoreCase))
+    public override List<string> GetHeaderNamesForSignature()
+    {
+        var headersList = new List<string>(NecessaryHeaders.Count);
+
+        // Check HTTP headers collection
+        foreach (var (Name, Necessity) in NecessaryHeaders)
+        {
+            if (string.Equals(Name, ResponseStatusHeaderName, StringComparison.OrdinalIgnoreCase))
+            {
+                headersList.Add(ResponseStatusHeaderName);
+            }
+            else if (string.Equals(Name, DigestHeadertName, StringComparison.OrdinalIgnoreCase))
+            {
+                headersList.Add(DigestHeadertName);
+            }
+            else
+            {
+                if (Headers.TryGetValue(Name, out string headerValue))
                 {
-                    sb.Append(hn)
-                        .Append(HttpMessageData.HeaderNameValueSeparator)
-                        .Append(StatusCode);
+                    headersList.Add(Name.ToLowerInvariant());
                 }
                 else
                 {
-                    if (!Headers.TryGetValue(hn, out string headerValue)
-                        && (additionalHeaders == null || !additionalHeaders.TryGetValue(hn, out headerValue)))
+                    if (Necessity == HeaderNecessity.Mandatory)
                     {
-                        throw new HeaderMissingException($"Can't find header '{hn}'");
-                    }
-
-                    sb.Append(hn)
-                        .Append(HeaderNameValueSeparator)
-                        .Append(headerValue);
-                }
-            }
-
-            return sb.ToString();
-
-        }
-
-        public override List<string> GetHeaderNamesForSignature()
-        {
-            var headersList = new List<string>(NecessaryHeaders.Count);
-
-            // Check HTTP headers collection
-            foreach (var (Name, Necessity) in NecessaryHeaders)
-            {
-                if (string.Equals(Name, ResponseStatusHeaderName, StringComparison.OrdinalIgnoreCase))
-                {
-                    headersList.Add(ResponseStatusHeaderName);
-                }
-                else if (string.Equals(Name, DigestHeadertName, StringComparison.OrdinalIgnoreCase))
-                {
-                    headersList.Add(DigestHeadertName);
-                }
-                else
-                {
-                    if (Headers.TryGetValue(Name, out string headerValue))
-                    {
-                        headersList.Add(Name.ToLowerInvariant());
-                    }
-                    else
-                    {
-                        if (Necessity == HeaderNecessity.Mandatory)
-                        {
-                            throw new HeaderMissingException($"Mandatory header missing {Name}");
-                        }
+                        throw new HeaderMissingException($"Mandatory header missing {Name}");
                     }
                 }
             }
-
-            return headersList;
         }
+
+        return headersList;
     }
 }
