@@ -20,18 +20,17 @@
  * SOFTWARE.
  */
 
-namespace TBC.OpenBanking.Jws
+namespace TBC.OpenBanking.Jws;
+
+using System;
+using System.Text;
+
+// Micro-optimization: Allow for de-virtualization (see https://github.com/dotnet/coreclr/pull/9230)
+internal sealed class UTF8EncodingSealed : UTF8Encoding
 {
-    using System;
-    using System.Text;
+    internal static readonly UTF8EncodingSealed Instance = new();
 
-    // Micro-optimization: Allow for de-virtualization (see https://github.com/dotnet/coreclr/pull/9230)
-    internal sealed class UTF8EncodingSealed : UTF8Encoding
-    {
-        internal static readonly UTF8EncodingSealed Instance = new UTF8EncodingSealed();
+    public UTF8EncodingSealed() : base(encoderShouldEmitUTF8Identifier: false, throwOnInvalidBytes: true) { }
 
-        public UTF8EncodingSealed() : base(encoderShouldEmitUTF8Identifier: false, throwOnInvalidBytes: true) { }
-
-        public override byte[] GetPreamble() => Array.Empty<byte>();
-    }
+    public override byte[] GetPreamble() => Array.Empty<byte>();
 }
